+++ R/calc_nreps2.R
@@ -99,7 +99,9 @@
 #' @param boot.R number of bootstrap resamples
 #' @param force.balanced logical flag to force the use of balanced sampling for
 #'        the algorithms on each instance
-#' #@param ncpus number of cores to use (under development.) #//DoParallel
+#' @param save.to.file logical flag: should the results be saved to a file
+#'        in the current working directory?
+#'
 #'
 #' @return a list object containing the following items:
 #' \itemize{
@@ -178,6 +180,23 @@
 #'     "\nphi_j =", my.reps$phi.est, "\nse    =", my.reps$se)
 #' }
 #'
+#'
+#' \dontrun{
+#' # Example for a 21-city TSP instance using 2 configurations of SANN
+#' algorithm1 <- list(FUN  = "my.SANN", alias = "algo1",
+#'                    Temp = 2000, budget = 10000)
+#' algorithm2 <- list(FUN  = "my.SANN", alias = "algo2",
+#'                    Temp = 4000, budget = 10000)
+#' instance <- list(FUN    = "TSP.dist",
+#'                  mydist = datasets::eurodist)
+#' my.reps  <- calc_nreps2(instance, algorithm1, algorithm2,
+#'                         se.max = 0.01, dif = "perc",
+#'                         method = "param", seed = 1234,
+#'                         nstart = 20)
+#' cat("n1j   =", my.reps$n1j, "\nn2j   =", my.reps$n2j,
+#'     "\nphi_j =", my.reps$phi.est, "\nse    =", my.reps$se)
+#' }
+#'
 #' @export
 
 # TESTED
@@ -191,8 +210,8 @@ calc_nreps2 <- function(instance,         # instance parameters
                         nmax   = 200,    # maximum allowed sample size
                         seed   = NULL,    # seed for PRNG
                         boot.R = 999,     # number of bootstrap resamples
-                        force.balanced = FALSE) # force balanced sampling
-#                        ncpus  = 1)       # number of cores to use #//DoParallel
+                        force.balanced = FALSE, # force balanced sampling
+                        save.to.file  = FALSE) # save results to tmp file
 {
 
   # ========== Error catching ========== #
@@ -210,36 +229,20 @@ calc_nreps2 <- function(instance,         # instance parameters
     nmax >= 2 * nstart,
     is.null(seed) || assertthat::is.count(seed),
     assertthat::is.count(boot.R), boot.R > 1,
-    is.logical(force.balanced), length(force.balanced) == 1)
-#    assertthat::is.count(ncpus)) #//DoParallel
+    is.logical(force.balanced), length(force.balanced) == 1,
+    is.logical(save.to.file), length(save.to.file) == 1)
   # ==================================== #
 
   # set PRNG seed
   if (is.null(seed)) {
+    if (!exists(".Random.seed")) stats::runif(1)
     seed <- .Random.seed #i.e., do not change anything
   } else{
     set.seed(seed)
   }
 
-
-  # # Set up doParallel       #//DoParallel
-  # local.cluster <- FALSE
-  # if (ncpus > 1){
-  #   cl.workers <- getDoParWorkers()
-  #   if (cl.workers < ncpus){
-  #     available.cores <- parallel::detectCores()
-  #     if (ncpus >= available.cores){
-  #       warning("ncpus too large, we only have ", available.cores, " cores. ",
-  #               "Using ", available.cores - 1, " cores.")
-  #       ncpus <- available.cores - 1
-  #     }
-  #     if (cl.workers < ncpus){
-  #       cl.calc_nreps2 <- parallel::makeCluster(ncpus)
-  #       doParallel::registerDoParallel(cl.calc_nreps2)
-  #       local.cluster <- TRUE
-  #     }
-  #   }
-  # }
+  # Echo some information for the user
+  cat("\nSampling algorithms on instance:", instance$alias)
 
   # generate initial samples
   n1j <- nstart # initial number of observations
@@ -278,6 +281,7 @@ calc_nreps2 <- function(instance,         # instance parameters
                   dif = dif, method = method, boot.R = boot.R)
   }
 
+  # Assemble output list
   output <- list(x1j     = x1j,
                  x2j     = x2j,
                  phi.est = SE$x.est,
@@ -287,22 +291,24 @@ calc_nreps2 <- function(instance,         # instance parameters
                  seed    = seed,
                  dif     = dif)
 
-  # unregister cluster if needed
-  #if (local.cluster) { stopCluster(cl.calc_nreps2) } #//DoParallel
+  # Save to file if required
+  if (save.to.file){
+    # Get folder
+    if(!dir.exists("./nreps_files")) dir.create("./nreps_files")
 
+    # Get a unique filename
+    filename <- paste0("./nreps_files/nreps_file_",
+                       gsub(":", "-", as.character(Sys.time())),
+                       "_",
+                       paste(sample(c(letters, LETTERS), size = 10),
+                             collapse = ""),
+                       ".rds")
+    filename <- gsub(" ", "-", filename)
+
+    # save output to file
+    saveRDS(output, file = filename)
+  }
+
+  # Return output
   return(output)
-
-
-  # if (method == "param"){
-  #   output <- nreps_param(instance = instance,
-  #                         algorithm1 = algorithm1, algorithm2 = algorithm2,
-  #                         se.max = se.max, dif = dif, nstart = nstart,
-  #                         nmax = nmax, seed = seed)
-  # } else if (method == "boot"){
-  #   output <- nreps_boot(instance = instance,
-  #                         algorithm1 = algorithm1, algorithm2 = algorithm2,
-  #                         se.max = se.max, dif = dif, nstart = nstart,
-  #                         nmax = nmax, seed = seed, boot.R = boot.R,
-  #                         ncpus = ncpus)
-  # }
 }