+++ NEWS.md
@@ -1,2 +1,19 @@
+# CAISER 0.3
+* `run_experiment()` can now be run in parallel using multiple cores.
+* `run_experiment()` and `calc_nreps2()` can now save results to files.
+
+# CAISER 0.2.4
+* `run_experiment()` now forces the use of all available instances if `power >= 1`.
+
+# CAISER 0.2.3
+* Improved plot and summary functions for `CAISErPowercurve` objects.
+* Added options to `calc_power_curve()` to determine the range of effect sizes to consider.
+
+# CAISER 0.2.2
+* Added new example and use case to `calc_nreps2()`
+
+# CAISEr 0.2.1
+* Minor fixes, particularly in printing function.
+
 # CAISEr 0.2.0
-* Initial release on CRAN
+* Initial release on CRAN.